backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-jose[cryptography]
bcrypt
python-dotenv
pyodbc
Requests